+++ mod_pow.h
@@ -6,7 +6,7 @@ typedef unsigned long long ull;
 
 int mod_pow_recursive(int x, long y, int m);
 
-int mod_pow(ull x, ull y, ull m);
+ull mod_pow(ull x, ull y, ull m);
 
 int mod_fibonachi(unsigned long n, int m);
